% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsvis_app.R
\name{mdsvis_app}
\alias{mdsvis_app}
\title{Launch shiny app for MDS projection visualization}
\usage{
mdsvis_app(preLoadDemoDataset = FALSE, maxUploadSize = 50 * 1024^2)
}
\arguments{
\item{preLoadDemoDataset}{if TRUE, pre-load the \emph{Krieg_Anti_PD_1} dataset}

\item{maxUploadSize}{maximum size for file upload (50MB by default)}
}
\value{
no return value
}
\description{
Launch shiny app for MDS projection visualization
}
\examples{
if (interactive()) {
  mdsvis_app()
}
}
