% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch shiny app for MDS projection visualization}
\usage{
run_app()
}
\value{
no return value
}
\description{
Launch shiny app for MDS projection visualization
}
\examples{
if (interactive()) {
  run_app()
}
}
