% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSvis-package.R
\docType{package}
\name{MDSvis}
\alias{MDSvis-package}
\alias{MDSvis}
\title{MDSvis package}
\description{
\code{MDSvis} provides a shiny application for interactive visualisation of
Multi Dimensional Scaling (MDS) results.
}
\seealso{
\link[CytoMDS:CytoMDS-package]{CytoMDS}

\link{run_app}
}
\author{
\strong{Maintainer}: Philippe Hauchamps \email{philippe.hauchamps@uclouvain.be} (\href{https://orcid.org/0000-0003-2865-1852}{ORCID})

Authors:
\itemize{
  \item Andrea Vicini \email{andrea.vicini@uclouvain.be} (\href{https://orcid.org/0000-0001-9438-6909}{ORCID})
  \item Laurent Gatto \email{laurent.gatto@uclouvain.be} (\href{https://orcid.org/0000-0002-1520-2268}{ORCID})
}

Other contributors:
\itemize{
  \item Shabnam Zaman \email{shabnam.zaman@uclouvain.be} (\href{https://orcid.org/0000-0001-5399-0130}{ORCID}) [contributor]
}

}
\keyword{internal}
