% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdsvis_app.R
\name{mdsvis_app}
\alias{mdsvis_app}
\title{Launch shiny app for MDS projection visualization}
\usage{
mdsvis_app(preLoadDemoDataset = FALSE)
}
\arguments{
\item{preLoadDemoDataset}{if TRUE, pre-load the \emph{Krieg_Anti_PD_1} dataset}
}
\value{
no return value
}
\description{
Launch shiny app for MDS projection visualization
}
\examples{
if (interactive()) {
  mdsvis_app()
}
}
